AccountService/requirements.txt
fastapi
pydantic
pyjwt
uvicorn
bcrypt
psycopg2-binary
sqlalchemy
databases